Example solution for issue 156: Adding files with the same name #1
The user can supply their own name for files to be nammed in the build directory.
This avoids having files overwrite themselves.
Perhaps this can be automated? If so, it will need to be taken in account for when the user wants to copy the same file to two different places.